Fix AG6 doctor visibility CI lint debt
