+++ crates/atm-daemon-bootstrap/Cargo.toml
@@ -12,4 +12,3 @@ homepage.workspace = true
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.3.1" }
 atm-runtime = { path = "../atm-runtime", version = "1.3.1" }
-atm-storage = { path = "../atm-storage", version = "1.3.1" }
