+++ crates/atm-daemon/Cargo.toml
@@ -33,7 +33,6 @@ atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "
 atm-daemon-client = { path = "../atm-daemon-client", version = "1.3.1" }
 atm-runtime = { path = "../atm-runtime", version = "1.3.1", features = ["test-utils"] }
 atm-runtime-test-support = { path = "../atm-runtime-test-support" }
-atm-storage-rusqlite = { path = "../atm-storage-rusqlite", version = "1.3.1" }
 sc-observability = { workspace = true, features = ["fault-injection"] }
 serial_test = "3"
 tempfile = "3"
