+++ crates/atm-runtime/src/lib.rs
@@ -11,6 +11,10 @@ mod legacy_storage_adapters;
 mod replay_store;
 mod sqlite_observability;
 
+pub use atm_storage::contract::{
+    AddPeerInterfaceCommand, AllowHostCommand, AllowedHostName, AllowedHostRow, PeerInterfaceKey,
+    PeerInterfaceKind, PeerInterfaceRow, UpdatePeerInterfaceCommand,
+};
 pub use composition::{
     RuntimeAssembly, RuntimeAssemblyInputs, assemble_default_runtime, assemble_sqlite_runtime,
     default_local_runtime, with_default_allowed_host_store,
