+++ crates/atm/Cargo.toml
@@ -26,7 +26,6 @@ atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "
 atm-daemon-bootstrap = { path = "../atm-daemon-bootstrap", version = "1.3.1" }
 atm-daemon-client = { path = "../atm-daemon-client", version = "1.3.1" }
 atm-runtime = { path = "../atm-runtime", version = "1.3.1" }
-atm-storage = { path = "../atm-storage", version = "1.3.1" }
 anyhow.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
 clap = { version = "4", features = ["derive"] }
